src/layers/pip_modules/requirements.txt
boto3==1.9.100
pyyaml==5.1
requests==2.21.0
feedparser==5.2.1
BeautifulSoup4
